server/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.2
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
django-jazzmin==2.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
Pillow==8.4.0
pipenv==2021.5.29
platformdirs==2.2.0
PyJWT==2.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.7.2
virtualenv-clone==0.5.6
